Inform user if the config name has spaces stripped out
If the provided configuration name contains spaces, those are stripped
out for convenience when interacting with systemd and systemctl
operations.  If this happens, inform the user as long as the default
configuration name is not used.

The rationale of not informing if the default config name (OpenVPN
Cloud) is used, is that the user didn't set this value so this change
doesn't appear as an unexpected behaviour.

Signed-off-by: David Sommerseth <davids@openvpn.net>