Add a config name duplication check for configmgr imports
When importing configurations into the OpenVPN 3 Configuration Manager,
it is possible to have non-unique file names.  However, this will not
work well with the openvpn3-session@.service unit file, as it needs a
unique configuration name to pick the right configuration.

This adds a simple duplication check and will report an error if the
configuration name is already found.

Signed-off-by: David Sommerseth <davids@openvpn.net>