Move the init of ConfigImport object before downloading profile
The init phase of the ConfigImport object does the config name
duplication check.  By moving the init before the profile download we
will bail out earlier and quicker with an error in case of non-unique
configuration names and we don't issue any remote operations until we
are ready to process the retrieved data locally.

Signed-off-by: David Sommerseth <davids@openvpn.net>